# api/swagger/swagger.yaml
definitions:
  api_v1.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api_v1.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  dto.CreateCategoryDTO:
    properties:
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dto.UpdateCategoryDTO:
    properties:
      name:
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  models.Account:
    properties:
      account_type:
        $ref: '#/definitions/models.AccountType'
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      update_at:
        type: string
      user_id:
        type: integer
      username:
        minLength: 3
        type: string
    required:
    - username
    type: object
  models.AccountType:
    enum:
    - admin
    type: string
    x-enum-varnames:
    - Admin
  models.Category:
    properties:
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      update_at:
        type: string
    type: object
  models.Product:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      create_at:
        type: string
      delete_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      update_at:
        type: string
    type: object
...
  models.Ticket:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      account_id:
        type: integer
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      ticket_products:
        items:
          $ref: '#/definitions/models.TicketProduct'
        type: array
      ticket_status:
        $ref: '#/definitions/models.TicketStatus'
      total:
        type: number
      update_at:
        type: string
    type: object
  models.TicketProduct:
    properties:
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/models.Product'
      quantity:
        type: integer
      ticket_id:
        type: integer
      update_at:
        type: string
    type: object
  models.TicketStatus:
    enum:
    - open
    type: string
    x-enum-varnames:
    - TicketOpen
  models.User:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      create_at:
        type: string
      delete_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      update_at:
        type: string
    type: object
  models_errors.ErrorResponse:
    properties:
      description:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Point Of Sale - Total Tools
  title: Point Of Sale API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      parameters:
      - description: credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api_v1.Login'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_v1.LoginResponse'
      tags:
      - Auth
  /api/v1/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Category
    post:
      consumes:
      - application/json
      parameters:
      - description: Category DTO
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryDTO'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - ApiKeyAuth: []
      tags:
      - Category
  /api/v1/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      tags:
      - Category
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
      security:
      - ApiKeyAuth: []
      tags:
      - Category
    put:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category DTO
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryDTO'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      tags:
      - Category
...
  /api/v1/payments/tickets/{ticketId}:
    post:
      parameters:
      - description: Ticket ID
        in: path
        name: ticketId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Payments
...
  /api/v1/products/{productId}/categories/{categoryId}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Product Category
    post:
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Product Category
  /api/v1/products/categories/{idCategory}:
    get:
      parameters:
      - description: Category ID
        in: path
        name: idCategory
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      security:
      - ApiKeyAuth: []
      tags:
      - Product
...
  /api/v1/tickets/{id}:
    delete:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
    get:
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/models_errors.ErrorResponse'
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
      security:
      - ApiKeyAuth: []
      tags:
      - Ticket
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"